feat: RSP1A implementation using new API
