+++ src/spectre_core/receivers/plugins/_rsp1a.py
@@ -3,21 +3,167 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from dataclasses import dataclass
-from typing import Optional
+from typing import Optional, Callable
 from functools import partial
 
-from ._receiver_names import ReceiverName
-from .gr._gr_rsp1a import swept_center_frequency, fixed_center_frequency
-from .gr._base import capture
-from ._sdrplay_receiver import (
-    get_pvalidator_fixed_center_frequency,
-    get_pvalidator_swept_center_frequency,
-    get_capture_template_fixed_center_frequency,
-    get_capture_template_swept_center_frequency,
+from spectre_core.capture_configs import (
+    CaptureTemplate,
+    CaptureMode,
+    Parameters,
+    Bound,
+    PName,
+    OneOf,
+    get_base_capture_template,
+    get_base_ptemplate,
+    validate_fixed_center_frequency,
+    validate_swept_center_frequency,
 )
+
+from ._receiver_names import ReceiverName
+from ._rsp1a_gr import swept_center_frequency, fixed_center_frequency
+from ._receiver_names import ReceiverName
+from ._gr import capture
+from .._receiver import Specs, SpecName
+
+
 from .._receiver import SpecName, Receiver
 from .._register import register_receiver
-from ._receiver_names import ReceiverName
+
+
+def make_pvalidator_fixed_center_frequency(
+    specs: Specs,
+) -> Callable[[Parameters], None]:
+    def pvalidator(parameters: Parameters) -> None:
+        validate_fixed_center_frequency(parameters)
+
+    return pvalidator
+
+
+def make_pvalidator_swept_center_frequency(
+    specs: Specs,
+) -> Callable[[Parameters], None]:
+    def pvalidator(parameters: Parameters) -> None:
+        validate_swept_center_frequency(
+            parameters, specs.get(SpecName.API_RETUNING_LATENCY)
+        )
+
+    return pvalidator
+
+
+def make_capture_template_fixed_center_frequency(
+    specs: Specs,
+) -> CaptureTemplate:
+
+    capture_template = get_base_capture_template(CaptureMode.FIXED_CENTER_FREQUENCY)
+    capture_template.add_ptemplate(get_base_ptemplate(PName.BANDWIDTH))
+    capture_template.add_ptemplate(get_base_ptemplate(PName.IF_GAIN))
+    capture_template.add_ptemplate(get_base_ptemplate(PName.RF_GAIN))
+
+    capture_template.set_defaults(
+        (PName.BATCH_SIZE, 3.0),
+        (PName.CENTER_FREQUENCY, 95800000),
+        (PName.SAMPLE_RATE, 600000),
+        (PName.BANDWIDTH, 600000),
+        (PName.WINDOW_HOP, 512),
+        (PName.WINDOW_SIZE, 1024),
+        (PName.WINDOW_TYPE, "blackman"),
+        (PName.RF_GAIN, -30),
+        (PName.IF_GAIN, -30),
+    )
+
+    capture_template.add_pconstraint(
+        PName.CENTER_FREQUENCY,
+        [
+            Bound(
+                lower_bound=specs.get(SpecName.FREQUENCY_LOWER_BOUND),
+                upper_bound=specs.get(SpecName.FREQUENCY_UPPER_BOUND),
+            )
+        ],
+    )
+    capture_template.add_pconstraint(
+        PName.SAMPLE_RATE,
+        [
+            Bound(
+                lower_bound=specs.get(SpecName.SAMPLE_RATE_LOWER_BOUND),
+                upper_bound=specs.get(SpecName.SAMPLE_RATE_UPPER_BOUND),
+            )
+        ],
+    )
+    capture_template.add_pconstraint(
+        PName.BANDWIDTH, [OneOf(specs.get(SpecName.BANDWIDTH_OPTIONS))]
+    )
+    capture_template.add_pconstraint(
+        PName.IF_GAIN,
+        [Bound(upper_bound=specs.get(SpecName.IF_GAIN_UPPER_BOUND))],
+    )
+    capture_template.add_pconstraint(
+        PName.RF_GAIN,
+        [Bound(upper_bound=specs.get(SpecName.RF_GAIN_UPPER_BOUND))],
+    )
+    return capture_template
+
+
+def make_capture_template_swept_center_frequency(specs: Specs) -> CaptureTemplate:
+
+    capture_template = get_base_capture_template(CaptureMode.SWEPT_CENTER_FREQUENCY)
+    capture_template.add_ptemplate(get_base_ptemplate(PName.BANDWIDTH))
+    capture_template.add_ptemplate(get_base_ptemplate(PName.IF_GAIN))
+    capture_template.add_ptemplate(get_base_ptemplate(PName.RF_GAIN))
+
+    capture_template.set_defaults(
+        (PName.BATCH_SIZE, 4.0),
+        (PName.MIN_FREQUENCY, 95000000),
+        (PName.MAX_FREQUENCY, 100000000),
+        (PName.SAMPLES_PER_STEP, 80000),
+        (PName.FREQUENCY_STEP, 1536000),
+        (PName.SAMPLE_RATE, 1536000),
+        (PName.BANDWIDTH, 1536000),
+        (PName.WINDOW_HOP, 512),
+        (PName.WINDOW_SIZE, 1024),
+        (PName.WINDOW_TYPE, "blackman"),
+        (PName.RF_GAIN, -30),
+        (PName.IF_GAIN, -30),
+    )
+
+    capture_template.add_pconstraint(
+        PName.MIN_FREQUENCY,
+        [
+            Bound(
+                lower_bound=specs.get(SpecName.FREQUENCY_LOWER_BOUND),
+                upper_bound=specs.get(SpecName.FREQUENCY_UPPER_BOUND),
+            )
+        ],
+    )
+    capture_template.add_pconstraint(
+        PName.MAX_FREQUENCY,
+        [
+            Bound(
+                lower_bound=specs.get(SpecName.FREQUENCY_LOWER_BOUND),
+                upper_bound=specs.get(SpecName.FREQUENCY_UPPER_BOUND),
+            )
+        ],
+    )
+    capture_template.add_pconstraint(
+        PName.SAMPLE_RATE,
+        [
+            Bound(
+                lower_bound=specs.get(SpecName.SAMPLE_RATE_LOWER_BOUND),
+                upper_bound=specs.get(SpecName.SAMPLE_RATE_UPPER_BOUND),
+            )
+        ],
+    )
+    capture_template.add_pconstraint(
+        PName.BANDWIDTH, [OneOf(specs.get(SpecName.BANDWIDTH_OPTIONS))]
+    )
+    capture_template.add_pconstraint(
+        PName.IF_GAIN,
+        [Bound(upper_bound=specs.get(SpecName.IF_GAIN_UPPER_BOUND))],
+    )
+    capture_template.add_pconstraint(
+        PName.RF_GAIN,
+        [Bound(upper_bound=specs.get(SpecName.RF_GAIN_UPPER_BOUND))],
+    )
+    return capture_template
 
 
 @dataclass(frozen=True)
@@ -50,14 +196,14 @@ class RSP1A(Receiver):
         self.add_mode(
             Mode.FIXED_CENTER_FREQUENCY,
             partial(capture, top_block_cls=fixed_center_frequency),
-            get_capture_template_fixed_center_frequency(self.specs),
-            get_pvalidator_fixed_center_frequency(self.specs),
+            make_capture_template_fixed_center_frequency(self.specs),
+            make_pvalidator_fixed_center_frequency(self.specs),
         )
         self.add_mode(
             Mode.SWEPT_CENTER_FREQUENCY,
             partial(
                 capture, top_block_cls=swept_center_frequency, max_noutput_items=1024
             ),
-            get_capture_template_swept_center_frequency(self.specs),
-            get_pvalidator_swept_center_frequency(self.specs),
+            make_capture_template_swept_center_frequency(self.specs),
+            make_pvalidator_swept_center_frequency(self.specs),
         )
