+++ src/spectre_core/receivers/plugins/_rsp1a_gr.py
@@ -1,6 +1,6 @@
-#
-# RSP1A receiver top blocks
-#
+# SPDX-FileCopyrightText: © 2024-2025 Jimmy Fitzpatrick <jcfitzpatrick12@gmail.com>
+# This file is part of SPECTRE
+# SPDX-License-Identifier: GPL-3.0-or-later
 
 
 from spectre_core.capture_configs import Parameters, PName
